change routing structure and added home page
