+++ src/components/LoginButton/LoginButton.modules.css
@@ -1,4 +1,5 @@
 /* CSS */
+/*
 button {
   background-color: #282634;
   border: 1px solid transparent;
@@ -7,7 +8,7 @@ button {
   color: #FFFFFF;
   cursor: pointer;
   flex: 0 0 auto;
-  font-family: "Inter var",ui-sans-serif,system-ui,-apple-system,system-ui,"Segoe UI",Roboto,"Helvetica Neue",Arial,"Noto Sans",sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol","Noto Color Emoji";
+  font-family: "Inter var", ui-sans-serif, system-ui, -apple-system, system-ui, "Segoe UI", Roboto, "Helvetica Neue", Arial, "Noto Sans", sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji";
   font-size: 1rem;
   font-weight: 600;
   line-height: 1.5rem;
@@ -17,14 +18,14 @@ button {
   text-decoration: none #6B7280 solid;
   text-decoration-thickness: auto;
   transition-duration: .2s;
-  transition-property: background-color,border-color,color,fill,stroke;
+  transition-property: background-color, border-color, color, fill, stroke;
   transition-timing-function: cubic-bezier(.4, 0, 0.2, 1);
   user-select: none;
   -webkit-user-select: none;
   touch-action: manipulation;
 }
 
-button:hover {
+ button:hover {
   background-color: #5f92b7;
 }
 
@@ -33,5 +34,4 @@ button:focus {
   outline: 2px solid transparent;
   outline-offset: 2px;
 }
-
-
+*/
